+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Fix some potention NPE's
 - Fix native libraries without a classifier download failing
+- Fix legacy java fixer not being installed in older packs